+++ lib/components/dropdown/dropdown.dart
@@ -2,10 +2,10 @@ part of bs_dropdown;
 
 @Directive (selector: "bs-dropdown, .dropdown",
     host: const {"[class.dropdown]" : "true", "[class.open]" : "isOpen"})
-class Dropdown implements OnInit, OnDestroy {
+class BsDropdownDirective implements OnInit, OnDestroy {
   ElementRef elementRef;
 
-  Dropdown(this.elementRef);
+  BsDropdownDirective(this.elementRef);
 
   /// if `true` `dropdown-menu` content will be appended to the body. This is useful when
   /// the dropdown button is inside a div with `overflow: hidden`, and the menu would
@@ -62,7 +62,7 @@ class Dropdown implements OnInit, OnDestroy {
   @Output() EventEmitter isOpenChange = new EventEmitter();
 
   /// sets the element that will fire the toggle of the dropdown
-  set dropDownToggle(DropdownToggle dropdownToggle) {
+  set dropDownToggle(BsDropdownToggleDirective dropdownToggle) {
     // init toggle element
     toggleEl = dropdownToggle.elementRef;
   }
@@ -83,7 +83,7 @@ class Dropdown implements OnInit, OnDestroy {
   }
 
   /// sets the element that will be showed by the dropdown
-  set dropDownMenu(DropdownMenu dropdownMenu) {
+  set dropDownMenu(BsDropdownMenuDirective dropdownMenu) {
     // init drop down menu
     menuEl = dropdownMenu.elementRef;
     if (dropdownAppendToBody) {